Unify tap and long press playground screen
