+++ android/playground/discover/src/main/java/dev/jasonpearson/automobile/discover/DiscoverScreen.kt
@@ -34,7 +34,6 @@ fun DiscoverVideoScreen(
     val tabs: Map<String, @Composable () -> Unit> =
         mapOf(
             "Tap" to { TapScreen() },
-            "Long Press" to { LongPressScreen() },
             "Swipe" to { SwipeScreen() },
             "Media" to { VideoListScreen(onNavigateToVideoPlayer = onNavigateToVideoPlayer) },
             "Text" to { InputTextScreen() },
