+++ test/features/observe/PredictionAnalyzer.test.ts
@@ -27,8 +27,8 @@ const baseObservation: ObserveResult = {
     hierarchy: {
       node: {
         $: {
-          bounds: "[0,0][1,1]",
-          text: "Settings",
+          "bounds": "[0,0][1,1]",
+          "text": "Settings",
           "resource-id": "id/settings"
         }
       }
@@ -70,8 +70,8 @@ describe("PredictionAnalyzer", () => {
         hierarchy: {
           node: {
             $: {
-              bounds: "[0,0][1,1]",
-              text: "Settings",
+              "bounds": "[0,0][1,1]",
+              "text": "Settings",
               "resource-id": "id/settings",
               "content-desc": "Settings"
             }
